refactor http query tests and improve rest client error handling
